refactor(mysql): embed Extractor from BaseExtractor
